requirements.txt
requests
yfinance
numpy
pandas
python-dotenv